+++ Demo/VFLRunner/Sources/ViewController.swift
@@ -18,8 +18,7 @@ class ViewController: UIViewController {
     self.title = "VFL Demo"
     
     let tableVw = UITableView(frame: .zero, style: .plain)
-    vfl
-      .add(subview: tableVw, name: "tableVw")
+    vfl.add(subview: tableVw, name: "tableVw")
     tableVw.dataSource = self
     tableVw.delegate = self
     tableVw.register(UITableViewCell.self, forCellReuseIdentifier: Self.cellId)
@@ -28,7 +27,8 @@ class ViewController: UIViewController {
   override func viewSafeAreaInsetsDidChange() {
     super.viewSafeAreaInsetsDidChange()
     vfl
-      .replaceConstraints(
+      .removeAllConstraints()
+      .appendConstraints(
         metrics: view.safeAreaInsets.metrics,
         formats: [
           "V:|-(top)-[tableVw]-(bottom)-|",