+++ Sources/VFL/VFL.swift
@@ -13,7 +13,10 @@ public typealias VFLView = UIView
 public class VFL {
   private(set) var views: [String: VFLView] = [:]
   private(set) var parentView: VFLView?
+  private(set) var options: NSLayoutConstraint.FormatOptions = []
+  private(set) var metrics: [String: CGFloat] = [:]
   private(set) var constraints: [NSLayoutConstraint] = []
+  private(set) var formats:[String: [String]] = [:]
   
   public init(_ view: VFLView? = nil) {
     parentView = view
@@ -41,12 +44,49 @@ public class VFL {
   }
   
   @discardableResult
-  public func appendConstraints(
-    options: NSLayoutConstraint.FormatOptions = [],
-    metrics: [String: CGFloat]? = nil,
-    formats: [String]
-  ) -> VFL {
-    appendConstraints(formats.flatMap {
+  public func addOptions(_ options: NSLayoutConstraint.FormatOptions) -> VFL {
+    self.options.formUnion(options)
+    return self
+  }
+  
+  @discardableResult
+  public func removeAllOptions() -> VFL {
+    self.options = []
+    return self
+  }
+
+  
+  @discardableResult
+  public func addMetrics(_ metrics: [String: CGFloat]) -> VFL {
+    self.metrics.merge(metrics) { $1 }
+    return self
+  }
+
+  @discardableResult
+  public func removeAllMetrics() -> VFL {
+    self.metrics = [:]
+    return self
+  }
+  
+  @discardableResult
+  public func storeConstraints(formats: [String], name: String) -> VFL {
+    self.formats[name] = formats
+    return self
+  }
+  
+  @discardableResult
+  public func applyConstraintsNamed(name: String) -> VFL {
+    guard let formats = self.formats[name] else {
+      assertionFailure("No cached format found with name \(name)")
+      return self
+    }
+    applyConstraints(formats: formats)
+    return self
+  }
+
+  @discardableResult
+  public func applyConstraints(formats: [String]) -> VFL {
+    applyConstraints(constraints: formats.flatMap {
       NSLayoutConstraint.constraints(
         withVisualFormat: $0,
         options: options,
@@ -58,7 +98,7 @@ public class VFL {
   }
   
   @discardableResult
-  public func appendConstraints(_ constraints: [NSLayoutConstraint]) -> VFL {
+  public func applyConstraints(constraints: [NSLayoutConstraint]) -> VFL {
     NSLayoutConstraint.activate(constraints)
     self.constraints.append(contentsOf: constraints)
     return self
@@ -68,6 +108,34 @@ public class VFL {
   public func removeAllConstraints() -> VFL {
     NSLayoutConstraint.deactivate(self.constraints)
     self.constraints = []
+    return self
+  }
+
+  @discardableResult
+  public func removeAll() -> VFL {
+    removeAllConstraints()
+    removeAllMetrics()
+    removeAllOptions()
+    return self
+  }
+}
+
+// Convenience Utils
+extension VFL {
+  @discardableResult
+  public func appendConstraints(
+    options: NSLayoutConstraint.FormatOptions = [],
+    metrics: [String: CGFloat] = [:],
+    formats: [String]
+  ) -> VFL {
+    addOptions(options)
+    addMetrics(metrics)
+    applyConstraints(formats: formats)
+    return self
+  }
+  
+  public func appendConstraints(_ constraints: [NSLayoutConstraint]) -> VFL {
+    applyConstraints(constraints: constraints)
     return self
   }
 }